add feature definition for the tray
it contains a first scenario, but something seems broken with recent
rails. for some reasons the user model is not updated or updated wrong.
this is not a bug in the TrayPosition class, but somewhere in Asset,
acts_as_asset or User, not sure, what exactly goes wrong here.

comiting anyway, to have these tests in the repo