add root_frame accessor to presentation
the root frame should not be considered by any view, except the list of
possible parents for new frames.

unfortunately, for now, the frames association also contains the root
frame. this needs to be considered when working with a presentation’s
frames. (e.g. presentation.frames.count is to large, also while
iterating, one must check for the root frame)