add pending update spec for frames
I’m not exactly sure, how to handle this one, but I commit anyway. I
don’t know of any way to make the update fail. When using an unknown
attribute, it just crashes (which is a good thing). IMHO this can only
happen, when a validation fails. The only validations I know of, check
for presence, which is the case, since we are only updating an existing
record.