make navigation head work like it did in xmendel
The presentation title is a link to the first slide. this acts as a
title slide and is not shown in the outline.

then the author is parsed from the fassets. this still needs some work,
because for now only the caption is shown, which probably isn’t, what we
want in this case