Make the necessary changes to reflect the new creation and edit structure of fassets_core
Make the changes to integrate presentations in the add_asset and edit_box structure.
This includes moving the frame creation to the frames editor to clean up the editing of
presentations. This makes presentations act as all other assets.
Introduce an "Editor" button to move to the editing portion of a presentations. This better
reflects that the frame editor is the part where a presentations is really edited.