Add the wysiwyg-editor part for the new markup commands.
Add some new buttons to the editor to insert definitions, examples, box and foreign.
Style the new commands through css. For now this is only done for the default template.
Add an extension to allow the insertion and editing of tables.
Add a frames_helper to easily convert between the new formats.

Fix:
Stay in the wysiwyg editor when a new frame is selected from the sidebar.