Add the \cite command and allow formating inside the boxes.
Add the new \cite["bibKey"]{"quote"} command. This resolves the citation information from
bibtex files. For this a new directory app/bibtex/ is used. The converter searches in all
the files in that directory for the specified key. At the moment the formating is:
authors (year)

Allow the use of ordered/unordered lists, blank lines and other formating inside the content
of definitions, examples, boxes and citations.