Add new kind of wysiwyg-editor based on in-place editing
To circumvent all the restrictions of normal iframe based editors. Replace it with an in-place editing version.
It is based on the jquery-plugin jeditable. Now every text element is editable only in predefined places.
For example the definition title and content. This ensures the resulting html is correct and can be convert to
markdown. Paragraphs can be inserted before or after every element by clicking at that area.

A nice bonus is that all elements can now be sorted inside of a slot through the use of .sortable().
That is important because there is no way to mark a position in a slot for the insertion of elements, ie examples.
One problem is that, for now, the editing is done in markdown, which is then converted to html. This might
be fixable through the use of a lightweight wysiwyg-editor for the textareas.

In this commit the \cite commands is not yet integrated into the wysiwyg editor. This will be done later.