New WYSIWYG editor on Mercury Editor basis.
Add a completely new version of the wysiwyg editor. This version uses the mercury editor as basis.
Mercury is a HTML wysiwyg editor on HTML5 basis that solves most problems of other available editors
like event handling, content styling and extendability. The new version uses a new layout that minimizes
the space occupied by forms and buttons and groups them directly in the mercury toolbar.
A new preview window can be opened from the editor that shows the frame in the presentation mode and
updates on save and every frame change.
Slots can now be resized and the content of two slots can by switched with the click of a button.