Improve resizing and slot switching in the markup editor
Change the resizing and slot switching to a method approach to remove a lot of repetition.
Resizing and switching consists of a few operations on slots that repeat themself a lot.
For this three functions are implemented setSlotSize(), positionSlot() and switchSlots().
This makes the code more readable and removes bugs.