Improve the rearrangement of slots.
The new version of arrange_slots() ignores empty slots. This makes the template change to a template with fewer slots work  if the number of filled
slots is smaller or equal to the number of slots provided by the new template.
Slot sizes are reset to default values to ensure the new template is displayed correctly.