Introduce clipboard for slots.
This commit introduces a clipboard for slots that is shown left of the tray.
At the moment all slots that are not included in any frame template of the presentation
are added and can be dragged to slots to fill them with the content of that slot.

Additional fixes:
Improve the arrange_slots() function to include all filled slots,
including slots that are not in the current template. This allows the user to change
to a template with less slots and back without losing information or manually dragging
the slots from the clipboard back into the template.